requirements.txt
sentence-transformers
torch
numpy
transformers
accelerate
pandas
pyarrow
PyYAML
